+++ src/App.vue
@@ -4,54 +4,30 @@
       <h1 class="game-title">Simon</h1>
       <div class="round-counter">Round: {{ round }}</div>
       <div class="game">
-        <div class="game-field">
-          <div @click="registerClick"
-               :class="{'disabled': !active}"
-               class="field">
-            <div
-              ref="blue"
-              class="button blue"
-              data-color="blue"
-            ></div>
-            <div
-              ref="red"
-              class="button red"
-              data-color="red"
-            ></div>
-            <div
-              ref="yellow"
-              class="button yellow"
-              data-color="yellow"
-            ></div>
-            <div
-              ref="green"
-              class="button green"
-              data-color="green"
-            ></div>
-          </div>
-          <div
-            v-if="!active && round !==0 "
-            class="game-notify">
-            Sorry. You have failed :(
-            Save Results?
-            <button
-              v-if="!active && round !== 0"
-              @click="saveResult"
-              class="start-button">Save Results</button>
-          </div>
-        </div>
+        <game-field
+          class="game-field"
+          :difficulty = "difficulty"
+          :active = "active"
+          @next-round = "onNextRound"
+          @game-start = "onGameStart"
+          @game-end = "onGameEnd"
+        />
+        <game-notification
+          class="game-notification"
+          :active = "active"
+          :difficulty ="difficulty"
+          :round = "round"
+        />
         <div class="game-settings">
           <game-settings
             :active = "active"
-            @change-difficulty = "($difficulty) => this.difficulty = $difficulty"
-          />
-          <start-button
-            :active = "active"
-            @click-start = "startButtonHandler"
+            @change-difficulty = "difficultyChangeHandler"
           />
         </div>
       </div>
-      <game-scoreboard />
+      <div class="game-scoreboard">
+        <game-scoreboard />
+      </div>
       <div class="footer">
         <div class="footer-text">Created by <a href="https://github.com/nikitavik">Nikita Kornilov</a></div>
       </div>
@@ -60,124 +36,42 @@
 </template>
 
 <script>
-
-import { playSound } from './audio-api'
-import { getScores, postScores } from './api'
 import GameScoreboard from './components/GameScoreboard'
-import StartButton from './components/StartButton'
 import GameSettings from './components/GameSettings'
+import GameField from './components/GameField'
+import GameNotification from './components/GameNotification'
+
 export default {
   name: 'App',
   components: {
+    GameNotification,
+    GameField,
     GameSettings,
-    StartButton,
     GameScoreboard
   },
   data () {
     return {
       difficulty: 'normal',
       active: false,
-      pattern: [],
       round: 0
     }
   },
-  created () {
-    window.p = postScores
-    window.g = getScores
-  },
   methods: {
-    // Starts and end the game with button
-    startButtonHandler () {
-      if (!this.active) {
-        this.gameStart()
-      } else {
-        this.gameEnd()
-      }
-    },
-    // Function that plays pattern for each round
-    async playPattern (pattern, difficulty) {
-      let TIME = 1000
-      if (difficulty === 'easy') { TIME = 1500 }
-      if (difficulty === 'hard') { TIME = 400 }
-      for (const item of this.sequence) {
-        await new Promise(resolve => setTimeout(() => {
-          resolve()
-        }, TIME))
-        item.cb()
-        this.animateColor(item.colorName, TIME)
-      }
-    },
     // Game Starter
-    gameStart () {
-      console.warn(`Game: STARTED with difficulty ${this.difficulty.toUpperCase()} and pattern ${this.sequence}`)
-      this.active = true
-      this.sequence = []
+    onGameStart () {
+      console.warn(`Game: STARTED with difficulty ${this.difficulty.toUpperCase()}`)
       this.round = 0
-      this.newRound()
+      this.active = true
     },
-    // Game End
-    gameEnd () {
-      console.warn(`Game: ENDED with difficulty ${this.difficulty.toUpperCase()} and pattern ${this.sequence}`)
+    onGameEnd () {
+      console.warn(`Game: ENDED with difficulty ${this.difficulty.toUpperCase()}`)
       this.active = false
     },
-    // New round generator
-    newRound () {
+    onNextRound () {
       this.round = this.round + 1
-      this.addNewColor(this.randomColor())
-      this.playPattern(this.pattern, this.difficulty)
-      this.copy = this.sequence.slice(0)
     },
-    // Send click to game
-    registerClick (event) {
-      if (this.active && event.target.dataset.color) {
-        const color = event.target.dataset.color
-        this.checkInput(color)
-        this.animateColor(color)
-        playSound(color)
-      }
-    },
-    // Animate colors with Timeout
-    animateColor (color, time) {
-      const activeColor = this.$refs[color]
-      activeColor.classList.add('active')
-      setTimeout(() => { activeColor.classList.remove('active') }, time)
-    },
-    // Pushes new color to sequence
-    addNewColor (colorName) {
-      const newColor = {
-        colorName: colorName,
-        cb () {
-          playSound(colorName)
-        }
-      }
-      this.sequence.push(newColor)
-    },
-    // Check input with sequence
-    checkInput (color) {
-      const targetColor = this.copy.shift()
-      this.active = (targetColor.colorName === color)
-      this.checkLose()
-    },
-    // Check if input was right
-    checkLose () {
-      if (this.copy.length === 0 && this.active) {
-        this.newRound()
-      } else if (!this.active) {
-        this.gameEnd()
-      }
-    },
-    // Generate new random color name
-    randomColor () {
-      const seed = Math.floor((Math.random() * 4))
-      const colors = ['yellow', 'blue', 'red', 'green']
-      return colors[seed]
-    },
-    saveResult () {
-      const name = prompt('Tell me your name, please!')
-      postScores(name, this.round, this.difficulty).then(() => {
-        alert('Successfully saved')
-      }
-      )
+    difficultyChangeHandler (difficulty) {
+      this.difficulty = difficulty
     }
   }
 }
@@ -196,44 +90,7 @@ export default {
 input{
   cursor: pointer;
 }
-/*Game Field*/
-.field{
-  width: 100%;
-  height: 100%;
-}
-.button{
-  opacity: .5;
-  height: 300px;
-  width: 300px;
-  -webkit-border-radius: 150px 150px 150px 150px;
-  border-radius: 150px 150px 150px 150px;
-  position: absolute;
-  text-indent: 10000px;
-  cursor: pointer;
-}
-.disabled{
-  opacity: .1
-}
-.button.active{
-  opacity: 1;
-  border: 4px solid black;
-}
-.red{
-  clip: rect(0px, 300px, 150px, 150px);
-  background-color: red;
-}
-.blue{
-  background-color: blue;
-  clip: rect(0px, 150px, 150px, 0px);
-}
-.yellow{
-  background-color: yellow;
-  clip: rect(150px, 150px, 300px, 0px);
-}
-.green{
-  background-color: green;
-  clip: rect(150px,300px, 300px, 150px);
-}
+
 /*Main Styles*/
 *{
   box-sizing: border-box;
@@ -251,7 +108,7 @@ input{
   font-size: 3rem;
   line-height: 120%;
 }
-.game-notify{
+.game-notification{
   display: flex;
   flex-direction: column;
   align-items: center;
